Refactor of loader2 and config.
Encapsulated configuration data in an array to reduce global variable
use and increase visibility of use of configuration data.

Created an object to wrap loader2 logic to eliminate global variable
references.

Performed multiple method extraction refactorings on logic from loader2.

This is another set of changes where I won't be upset if the upstream
maintainer rejects my efforts. I wanted to understand what was going on
in the code and this was a great help in wrapping my head around the
business logic on the server side.