Make default search index configurable.
This small patch allows setting the default index name for searches in the
config.php file.

When using an elasticsearch server with a single logstash index (not
partitioned by day) and additional indexes that are not related to logstash
performance can be improved slightly by querying specific indexes rather than
`_all`.