Use a full path for importing the configuration.
The `include 'config.php';` directive can be problematic if you are hosting
Kibana inside a vhost that has other php components and config.php in the
default search path is not the config.php that belongs to Kibana.

The defaulthvior  will make sure that config.php is loaded from the same
directory reguardless of search path by providing a full-qualifed path to the
file.

Additionally the change allows the apache config, .htaccess or a bootstrap
script to specify an alternate location for the config file by setting an
environment variable. This is nice if you have the sort of Ops team that
punches you in the face when you mix configuration and code in the same
directory. :)