Written a basic multi-user access system
I've found some code in commit 4052967 for requiring a username and
password for access and ressurected it with some changes to allow
multiple users to have access to different subsets of the logs.
Each user has their own filter string which is used to configure
their access in the users.php file.