Crude user management page: admin.php
This page allows designated admin users to edit other users'
details and create new users. The interface is clunky and ugly,
there is almost certianly scope for injection attacks, but it does
work.

The users.php file must have write permissions for the apache user
in order to manipulate the users from the web interface.