Rewritten to store user credentials in elasticsearch
I've decided that elasticsearch probably is the place to keep
credentials; the fact that they are accessable to anyone with
access to port 9200 on the ES server is not too much of a problem
as these credentials only allow access to data stored there anyway.
The passwords are salted and hashed so people reusing passwords
are protected.

If use_auth is on but no users exist, a default user with username
admin, password pass and admin rights is created.

The code is currently very ugly; it needs all sorts of cleaning up
and refactoring, but does seem to work from quick testing.