Made my changes less ugly
I've tweaked the css properties to give me enough room in the nav
bar for a few basic navigation links. I've also added a navbar to
the admin page.

I've used jquery and bootstrap to improve the display of admin.
They seem to do something a bit odd with the toggle animation
which needs looking at, but it doesn't affect functionality.