Add file system based saved searches.
A quick and dirty hack to add the ability to save searches to files on the
file system. There are no permissions checks to see who can save, load or
delete a search. The same list of searches is seen by all users of the
instance.

- Adds "Load" and "Save" buttons to side bar.
- "Load" button retrieves a list of currently saved searches from server and
  displays a dialog that allows the user to choose a search to load into the
  current session or to delete.
- "Save" prompts for a name and sends the current search hash to the server
  for storage under that name. If the current search started from a saved
  search, the same name will be pre-populated in the save dialog.
- Adds restful GET, PUT and DELETE endpoints at /api/saved/:name for server
  side handling of storage/retrieval activities.

Related to issue #114 at https://github.com/rashidkpc/Kibana