Changing the config file should not dirty the repo.
The config file contains implementation details that may change from
one deployment to another, and changing these values should not dirty
the git repo.

This can be done by moving the config file to a sample file, and
ignoring the actual config file in git. This means that a new step
is added to the install process: copying the sample file to the actual
location before making any site-specific changes.